Cryptographic-API-Implementation/requirements.txt
#conda create --name cyber python=3.9
fastapi
pycryptodome
cryptography
uvicorn
